currency-selector implementation
