+++ src/app/app-routing.module.ts
@@ -8,6 +8,7 @@ import { OperationModule } from './client-domain-component/operations/operation.
 import { LoginModule } from './client-domain-component/auth/login.modules';
 import { AuthGuard } from './client-domain-component/auth/auth-guard.service';
 
+
 const appRoutes: Routes = [
     { path: '', redirectTo: 'bk', pathMatch: 'full' },
     { path: ':id', loadChildren: () => ClientDomainModule }
