+++ src/app/client-domain-component/operations/operation-routing.module.ts
@@ -1,5 +1,4 @@
 import { NgModule } from '@angular/core';
-
 import { Routes, RouterModule } from '@angular/router';
 import { AuthGuard } from '../auth/auth-guard.service';
 import { OperationsComponent } from './operations.component';
@@ -15,6 +14,9 @@ const operationRoutes: Routes = [
     imports: [
         RouterModule.forChild(operationRoutes)
     ],
+    providers:[ 
+   
+     ],
     exports: [RouterModule]
 })
 export class OperationRoutingModule {
