+++ src/app/client-domain-component/operations/premium/premium-actions/premium-actions.component.html
@@ -1,4 +1,4 @@
-<div class="row">
+<div class="row side-action">
 <div class="col-md-8">
 <div class="col-xs-3 col-md-8">
   <img src="../../assets/learn-more.png" class="img-responsive center-block" alt="Responsive image">
