+++ src/app/client-domain-component/operations/premium/premium-operations/premium-form/premium-form.component.ts
@@ -1,4 +1,6 @@
 import { Component, OnInit } from '@angular/core';
+//import{OperationService} from '../../../shared/operation.service';
+import{ICurrency} from'../../../shared/currency-selector/currency.model'
 
 @Component({
   selector: 'app-premium-form',
@@ -6,10 +8,18 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./premium-form.component.css']
 })
 export class PremiumFormComponent implements OnInit {
+  currencies:ICurrency[] 
 
   constructor() { }
 
   ngOnInit() {
+ //  this.currencies =  this.operationService.getCurrencyList()
+ //private operationService:OperationService
+  }
+  currency1 ={
+    symbol: "USD",
+    name: "US Dollars"
+              
   }
 
 }
