+++ src/app/client-domain-component/operations/premium/premium-operations/premium-operations.component.html
@@ -1,6 +1,11 @@
 <div class="row center-block">
 <p class="bg-primary">GET INTO THE OPERATIONS</p>
-<p>
-  premium-operations works!
-</p>
-</div>
+
+ <div class="row">
+    <div class="col-md-8">
+        <app-premium-form></app-premium-form>
+    </div>
+    <div class="col-md-4">
+        <app-premium-results></app-premium-results>
+    </div>
+</div>
