+++ src/app/client-domain-component/operations/premium/premium.component.html
@@ -1,9 +1,9 @@
 <div class="row">
-  <div class="col-md-3">
+  <div class="col-md-2">
     <app-premium-actions></app-premium-actions>
   </div>
 
-  <div class="col-md-9">
-     <app-premium-operations></app-premium-operations>
+  <div class="col-md-10">
+     <app-premium-operations></app-premium-operations> 
   </div>
 </div>
