+++ src/app/client-domain-component/operations/shared/currencies-list-directive.directive.spec.ts
