+++ src/app/client-domain-component/operations/shared/currency-selector/currency-selector.component.html
@@ -1,0 +1,3 @@
+ <li><a >{{currency.symbol}}-{{currency.name}}</a></li>   
+        
+
