+++ src/app/client-domain-component/operations/shared/currency-selector/currency-selector.component.spec.ts
