--- src/app/client-domain-component/shared/currency-selector/currency-selector.component.html
@@ -1,5 +1,0 @@
-<p>
-   
-
-  
-</p>
