--- src/app/client-domain-component/shared/currency-selector/currency-selector.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-currency-selector',
-  templateUrl: './currency-selector.component.html',
-  styleUrls: ['./currency-selector.component.css']
-})
-export class CurrencySelectorComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
